+++ src/attributes.js
@@ -315,6 +315,72 @@ function attachAttributeUpdater(Class, method, key, elementName, checker, finder
 }
 
 /**
+ * Attach an attribute remover method onto the provided class.
+ *
+ * @param {function} Class       - Target class.
+ * @param {string}   method      - Method name.
+ * @param {string}   key         - Key of the element's storage on instance.
+ * @param {string}   elementName - Name of target element for messages.
+ * @param {string}   checker     - Name of the checker method to use.
+ * @param {string}   [finder]    - Name of the finder method to use.
+ */
+function attachAttributeRemover(Class, method, key, elementName, checker, finder) {
+  const eventName = elementName === 'node' ? 'nodeAttributesUpdated' : 'edgeAttributesUpdated';
+
+  /**
+   * Remove the desired attribute for the given element (node or edge).
+   *
+   * Arity 2:
+   * @param  {any}    element - Target element.
+   * @param  {string} name    - Attribute's name.
+   *
+   * Arity 3 (only for edges):
+   * @param  {any}     source - Source element.
+   * @param  {any}     target - Target element.
+   * @param  {string}  name   - Attribute's name.
+   *
+   * @return {Graph}          - Returns itself for chaining.
+   *
+   * @throws {Error} - Will throw if too many arguments are provided.
+   * @throws {Error} - Will throw if any of the elements is not found.
+   */
+  Class.prototype[method] = function(element, name) {
+    if (arguments.length > 2) {
+      if (!finder)
+        throw new InvalidArgumentsGraphError(`Graph.${method}: too many arguments provided.`);
+
+      const source = element,
+            target = name;
+
+      name = arguments[2];
+
+      if (!this[checker](source, target))
+        throw new NotFoundGraphError(`Graph.${method}: could not find an edge for the given path ("${source}" - "${target}").`);
+
+      element = this[finder](source, target);
+    }
+
+    if (!this[checker](element))
+      throw new NotFoundGraphError(`Graph.${method}: could not find the "${element}" ${elementName} in the graph.`);
+
+    const data = this[key].get(element);
+
+    delete data.attributes[name];
+
+    // Emitting
+    this.emit(eventName, {
+      key: element,
+      type: 'remove',
+      meta: {
+        name
+      }
+    });
+
+    return this;
+  };
+}
+
+/**
  * Attach an attribute replacer method onto the provided class.
  *
  * @param {function} Class       - Target class.
@@ -478,6 +544,10 @@ const ATTRIBUTES_METHODS = [
   {
     name: element => `update${element}Attribute`,
     attacher: attachAttributeUpdater
+  },
+  {
+    name: element => `remove${element}Attribute`,
+    attacher: attachAttributeRemover
   },
   {
     name: element => `replace${element}Attributes`,